# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/noticias": {
            "get": {
                "tags": [
                    "Noticia"
                ],
                "summary": "Get a listing of the Noticias.",
                "description": "Get all Noticias",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Noticia"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Noticia"
                ],
                "summary": "Store a newly created Noticia in storage",
                "description": "Store Noticia",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Noticia that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Noticia"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Noticia"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/novedades": {
            "get": {
                "tags": [
                    "Novedades"
                ],
                "summary": "Get a listing of the Novedades.",
                "description": "Get all Novedades",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Novedades"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Novedades"
                ],
                "summary": "Store a newly created Novedades in storage",
                "description": "Store Novedades",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Novedades that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Novedades"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Novedades"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/clientes": {
            "get": {
                "tags": [
                    "cliente"
                ],
                "summary": "Get a listing of the clientes.",
                "description": "Get all clientes",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/cliente"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "cliente"
                ],
                "summary": "Store a newly created cliente in storage",
                "description": "Store cliente",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "cliente that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/cliente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/cliente"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/mensajes": {
            "get": {
                "tags": [
                    "mensaje"
                ],
                "summary": "Get a listing of the mensajes.",
                "description": "Get all mensajes",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/mensaje"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "mensaje"
                ],
                "summary": "Store a newly created mensaje in storage",
                "description": "Store mensaje",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "mensaje that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/mensaje"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/mensaje"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/personas": {
            "get": {
                "tags": [
                    "persona"
                ],
                "summary": "Get a listing of the personas.",
                "description": "Get all personas",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/persona"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "persona"
                ],
                "summary": "Store a newly created persona in storage",
                "description": "Store persona",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "persona that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/persona"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/persona"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Noticia": {
            "required": [
                "Titulo",
                "Descripcion",
                "UrlImage"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "Titulo": {
                    "description": "Titulo",
                    "type": "string"
                },
                "Descripcion": {
                    "description": "Descripcion",
                    "type": "string"
                },
                "UrlImage": {
                    "description": "UrlImage",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "Novedades": {
            "required": [
                "Titulo",
                "Descripcion",
                "Usuario",
                "Correo"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "Titulo": {
                    "description": "Titulo",
                    "type": "string"
                },
                "Descripcion": {
                    "description": "Descripcion",
                    "type": "string"
                },
                "Usuario": {
                    "description": "Usuario",
                    "type": "string"
                },
                "Correo": {
                    "description": "Correo",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "cliente": {
            "required": [
                ""
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "cedula": {
                    "description": "cedula",
                    "type": "string"
                },
                "nombre": {
                    "description": "nombre",
                    "type": "string"
                },
                "apellido": {
                    "description": "apellido",
                    "type": "string"
                },
                "edad": {
                    "description": "edad",
                    "type": "string"
                },
                "estatura": {
                    "description": "estatura",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "mensaje": {
            "required": [
                ""
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "mensaje": {
                    "description": "mensaje",
                    "type": "string"
                },
                "url": {
                    "description": "url",
                    "type": "string"
                },
                "data": {
                    "description": "data",
                    "type": "string"
                },
                "buttons": {
                    "description": "buttons",
                    "type": "string"
                },
                "schedule": {
                    "description": "schedule",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "persona": {
            "required": [
                ""
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "cedula": {
                    "description": "cedula",
                    "type": "integer",
                    "format": "int32"
                },
                "edad": {
                    "description": "edad",
                    "type": "integer",
                    "format": "int32"
                },
                "nombres": {
                    "description": "nombres",
                    "type": "string"
                },
                "apellidos": {
                    "description": "apellidos",
                    "type": "string"
                },
                "estatura": {
                    "description": "estatura",
                    "type": "string"
                },
                "selfie": {
                    "description": "selfie",
                    "type": "string",
                    "format": "binary"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        }
    }
}